fixed persist error where getting INSERT INTO measure (EQV, ndbno_id, LABEL, nutrient_id, QTY, VALUE) VALUES (?, ?, ?, ?, ?, ?)
;
**REAL COMMIT MESSAGE **
Fixed persist error where getting Foreign key exception by changing the persist cascade in Food for the meausurements array to .REMOVE only, added Cascade to Nutrient's
arrayList of measurements.  Fixed testentities so that the *Nutrient* Integer was not being set and did the same to the javascript object as both were setting the foreign key
nutrient_id where JPA is the one that sets the foreign key through its cascading persist.